+++ HydrogenEngine/Source/Platform/Vulkan/VulkanIndexBuffer.cpp
@@ -3,7 +3,7 @@
 
 using namespace Hydrogen;
 
-VulkanIndexBuffer::VulkanIndexBuffer(const std::shared_ptr<RenderContext>& renderContext, const std::vector<uint16_t>& indices)
+VulkanIndexBuffer::VulkanIndexBuffer(const std::shared_ptr<RenderContext>& renderContext, const std::vector<uint32_t>& indices)
 	: m_RenderContext(RenderContext::Get<VulkanRenderContext>(renderContext)),
 	VulkanBuffer(RenderContext::Get<VulkanRenderContext>(renderContext), sizeof(indices[0]) * indices.size(), VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_INDEX_BUFFER_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT)
 {
